fix(user): Cập nhật lại kiểu dữ liệu nhận vào từ Frontend ở các trường validate và chuyển đổi thành snake_case refactor: Chỉnh sửa lại logic sử dụng model, kèm tách biệt giữa domain và database entity

+++ internal/dto/userDTO/user_dto.go
@@ -9,7 +9,7 @@ type UserDTO struct {
 	Avatar      string `json:"avatar,omitempty"`
 	PhoneNumber string `json:"phoneNumber,omitempty"`
 	Address     string `json:"address,omitempty"`
-	Status      int    `json:"status"`
+	Status      int8   `json:"status"`
 	GoodPoint   int    `json:"goodPoint"`
 	Major       string `json:"major,omitempty"`
 }
